Move players component to player folder
