+++ src/components/page.component.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Board from './board/board.component';
-import Players from './players.component';
+import Players from './player/players.component';
 
 const Page = props => (
   <div className="container-fluid d-flex h-100 flex-column">
